chore(issue-44): resolves rerender on state-change
This is only a workaround. The issue still remains when using a loading component.
isResolving is only relevant when using a loading component, this workaround fixes at
least the disputed behavior when not using a loading component.

On the other hand this is totally fine, because when using a loading component, this
might be the desired behavior, that the loading component is shown on state change

Signed-off-by: Christian Sterzl <christian.sterzl@gmail.com>