fix: broken lazy loading resolved check
Resolving needs to be considered done once
the element is defined, not if the element
is undefined. This is a typo introduced in
41b69c718.

Also adapt the demo to use real lazy loading again.
The usage of .route="${import(…)}" means the component
is loaded by the time the <lit-route> is rendered,
not when the route actually becomes active.